requirements.txt
# Minimal dependencies for corpus conversion and text extraction
python-docx==1.1.2
pypdf==5.1.0

